LayoutTests/media/W3C/video/canPlayType/canPlayType_application_octet_stream_with_codecs_2-expected.txt

video.canPlayType() - can never play application/octet-stream with a codecs parameter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""

TEST COMPLETE
spec reference
